# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1//event/publish": {
            "post": {
                "description": "Publica uma nova mensagem no RabbitMQ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Publica um evento de mensagem no RabbitMQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Correlation Id",
                        "name": "x-correlation-id",
                        "in": "header"
                    },
                    {
                        "description": "Mensagem a ser publicada",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta de sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/event/GetRecentEvents": {
            "get": {
                "description": "Retorna todos os eventos dos últimos 5 minutos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Retorna todos os eventos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Event": {
            "type": "object",
            "properties": {
                "createdBy": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}